Move context menu handling to a new class
